refactor(react-query): move provider to libs folder
